+++ packages/mobile/babel.config.js
@@ -1,6 +1,4 @@
 module.exports = {
-  presets: [
-    'module:metro-react-native-babel-preset', 
-    '@babel/preset-typescript'
-  ],
-};
+  presets: ['module:metro-react-native-babel-preset'],
+  plugins: ['@babel/plugin-transform-flow-strip-types']
+}