+++ packages/mobile/src/components/ChannelTile/ChannelTile.test.tsx
@@ -20,6 +20,15 @@ describe('ChannelList component', () => {
 
     expect(toJSON()).toMatchInlineSnapshot(`
       <View
+        accessibilityState={
+          {
+            "busy": undefined,
+            "checked": undefined,
+            "disabled": undefined,
+            "expanded": undefined,
+            "selected": undefined,
+          }
+        }
         accessible={true}
         focusable={true}
         onClick={[Function]}
@@ -30,23 +39,24 @@ describe('ChannelList component', () => {
         onResponderTerminationRequest={[Function]}
         onStartShouldSetResponder={[Function]}
         style={
-          Object {
+          {
             "borderBottomColor": "#F0F0F0",
             "borderBottomWidth": 1,
             "padding": 16,
           }
         }
+        testID="channel_tile_general"
       >
         <View
           style={
-            Object {
+            {
               "flexDirection": "row",
             }
           }
         >
           <View
             style={
-              Object {
+              {
                 "alignItems": "center",
                 "flex": 1,
                 "paddingRight": 12,
@@ -55,7 +65,7 @@ describe('ChannelList component', () => {
           >
             <View
               style={
-                Object {
+                {
                   "alignItems": "center",
                   "backgroundColor": "#4C4C4C",
                   "borderRadius": 4,
@@ -70,8 +80,8 @@ describe('ChannelList component', () => {
                 fontSize={14}
                 horizontalTextAlign="left"
                 style={
-                  Array [
-                    Object {
+                  [
+                    {
                       "color": "#ffffff",
                       "fontFamily": "Rubik-Regular",
                       "fontSize": 14,
@@ -88,7 +98,7 @@ describe('ChannelList component', () => {
           </View>
           <View
             style={
-              Object {
+              {
                 "flex": 9,
                 "flexDirection": "column",
               }
@@ -96,14 +106,14 @@ describe('ChannelList component', () => {
           >
             <View
               style={
-                Object {
+                {
                   "flexDirection": "row",
                 }
               }
             >
               <View
                 style={
-                  Object {
+                  {
                     "flex": 8,
                   }
                 }
@@ -114,8 +124,8 @@ describe('ChannelList component', () => {
                   fontWeight="medium"
                   horizontalTextAlign="left"
                   style={
-                    Array [
-                      Object {
+                    [
+                      {
                         "color": "#000000",
                         "fontFamily": "Rubik-Medium",
                         "fontSize": 16,
@@ -132,7 +142,7 @@ describe('ChannelList component', () => {
               </View>
               <View
                 style={
-                  Object {
+                  {
                     "alignItems": "flex-end",
                     "flex": 4,
                   }
@@ -143,8 +153,8 @@ describe('ChannelList component', () => {
                   fontSize={14}
                   horizontalTextAlign="left"
                   style={
-                    Array [
-                      Object {
+                    [
+                      {
                         "color": "#999999",
                         "fontFamily": "Rubik-Regular",
                         "fontSize": 14,
@@ -161,7 +171,7 @@ describe('ChannelList component', () => {
             </View>
             <View
               style={
-                Object {
+                {
                   "flexDirection": "row",
                   "paddingTop": 3,
                 }
@@ -169,7 +179,7 @@ describe('ChannelList component', () => {
             >
               <View
                 style={
-                  Object {
+                  {
                     "flex": 10,
                   }
                 }
@@ -179,8 +189,8 @@ describe('ChannelList component', () => {
                   fontSize={14}
                   horizontalTextAlign="left"
                   style={
-                    Array [
-                      Object {
+                    [
+                      {
                         "color": "#7F7F7F",
                         "fontFamily": "Rubik-Regular",
                         "fontSize": 14,
@@ -196,7 +206,7 @@ describe('ChannelList component', () => {
               </View>
               <View
                 style={
-                  Object {
+                  {
                     "alignItems": "flex-end",
                     "flex": 2,
                   }
@@ -225,6 +235,15 @@ describe('ChannelList component', () => {
 
     expect(toJSON()).toMatchInlineSnapshot(`
       <View
+        accessibilityState={
+          {
+            "busy": undefined,
+            "checked": undefined,
+            "disabled": undefined,
+            "expanded": undefined,
+            "selected": undefined,
+          }
+        }
         accessible={true}
         focusable={true}
         onClick={[Function]}
@@ -235,23 +254,24 @@ describe('ChannelList component', () => {
         onResponderTerminationRequest={[Function]}
         onStartShouldSetResponder={[Function]}
         style={
-          Object {
+          {
             "borderBottomColor": "#F0F0F0",
             "borderBottomWidth": 1,
             "padding": 16,
           }
         }
+        testID="channel_tile_general"
       >
         <View
           style={
-            Object {
+            {
               "flexDirection": "row",
             }
           }
         >
           <View
             style={
-              Object {
+              {
                 "alignItems": "center",
                 "flex": 1,
                 "paddingRight": 12,
@@ -260,7 +280,7 @@ describe('ChannelList component', () => {
           >
             <View
               style={
-                Object {
+                {
                   "alignItems": "center",
                   "backgroundColor": "#4C4C4C",
                   "borderRadius": 4,
@@ -275,8 +295,8 @@ describe('ChannelList component', () => {
                 fontSize={14}
                 horizontalTextAlign="left"
                 style={
-                  Array [
-                    Object {
+                  [
+                    {
                       "color": "#ffffff",
                       "fontFamily": "Rubik-Regular",
                       "fontSize": 14,
@@ -293,7 +313,7 @@ describe('ChannelList component', () => {
           </View>
           <View
             style={
-              Object {
+              {
                 "flex": 9,
                 "flexDirection": "column",
               }
@@ -301,14 +321,14 @@ describe('ChannelList component', () => {
           >
             <View
               style={
-                Object {
+                {
                   "flexDirection": "row",
                 }
               }
             >
               <View
                 style={
-                  Object {
+                  {
                     "flex": 8,
                   }
                 }
@@ -319,8 +339,8 @@ describe('ChannelList component', () => {
                   fontWeight="medium"
                   horizontalTextAlign="left"
                   style={
-                    Array [
-                      Object {
+                    [
+                      {
                         "color": "#000000",
                         "fontFamily": "Rubik-Medium",
                         "fontSize": 16,
@@ -337,7 +357,7 @@ describe('ChannelList component', () => {
               </View>
               <View
                 style={
-                  Object {
+                  {
                     "alignItems": "flex-end",
                     "flex": 4,
                   }
@@ -348,8 +368,8 @@ describe('ChannelList component', () => {
                   fontSize={14}
                   horizontalTextAlign="left"
                   style={
-                    Array [
-                      Object {
+                    [
+                      {
                         "color": "#2373EA",
                         "fontFamily": "Rubik-Regular",
                         "fontSize": 14,
@@ -366,7 +386,7 @@ describe('ChannelList component', () => {
             </View>
             <View
               style={
-                Object {
+                {
                   "flexDirection": "row",
                   "paddingTop": 3,
                 }
@@ -374,7 +394,7 @@ describe('ChannelList component', () => {
             >
               <View
                 style={
-                  Object {
+                  {
                     "flex": 10,
                   }
                 }
@@ -384,8 +404,8 @@ describe('ChannelList component', () => {
                   fontSize={14}
                   horizontalTextAlign="left"
                   style={
-                    Array [
-                      Object {
+                    [
+                      {
                         "color": "#7F7F7F",
                         "fontFamily": "Rubik-Regular",
                         "fontSize": 14,
@@ -401,7 +421,7 @@ describe('ChannelList component', () => {
               </View>
               <View
                 style={
-                  Object {
+                  {
                     "alignItems": "flex-end",
                     "flex": 2,
                   }
@@ -409,7 +429,7 @@ describe('ChannelList component', () => {
               >
                 <View
                   style={
-                    Object {
+                    {
                       "alignItems": "center",
                       "backgroundColor": "#2373EA",
                       "borderRadius": 100,
@@ -425,8 +445,8 @@ describe('ChannelList component', () => {
                     fontWeight="medium"
                     horizontalTextAlign="left"
                     style={
-                      Array [
-                        Object {
+                      [
+                        {
                           "color": "#ffffff",
                           "fontFamily": "Rubik-Medium",
                           "fontSize": 12,