+++ packages/state-manager/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@1.2.0-alpha.24...@quiet/state-manager@1.2.0) (2023-04-18)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+# [1.2.0-alpha.24](/compare/@quiet/state-manager@1.2.0-alpha.23...@quiet/state-manager@1.2.0-alpha.24) (2023-04-17)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+# [1.2.0-alpha.23](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@1.2.0-alpha.22...@quiet/state-manager@1.2.0-alpha.23) (2023-04-14)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
 # [1.2.0-alpha.22](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@1.2.0-alpha.21...@quiet/state-manager@1.2.0-alpha.22) (2023-04-12)
 
 **Note:** Version bump only for package @quiet/state-manager